backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
numpy==2.2.1
langgraph==0.2.60
langchain-core==0.3.28
python-dotenv==1.0.1
httpx==0.28.1
